Reorg CAPI init, and add constructor for ClusterConfig with defaults
