+++ pkg/capi/capi_init_test.go
