+++ pkg/capi/cluster_lifecycle.go
@@ -4,9 +4,6 @@ package capi
 
 import (
 	"fmt"
-	os2 "github.com/verrazzano/verrazzano/pkg/os"
-	"go.uber.org/zap"
-
 	clusterapi "sigs.k8s.io/cluster-api/cmd/clusterctl/client"
 )
 
@@ -51,6 +48,16 @@ type ClusterLifeCycleManager interface {
 	Destroy() error
 }
 
+//NewClusterConfig Creates a new ClusterConfig with defaults
+func NewClusterConfig() ClusterConfig {
+	return ClusterConfig{
+		ClusterName:    bootstrapClusterName,
+		Type:           OCNEClusterType,
+		ContainerImage: getDefaultBoostrapImage(OCNEClusterType),
+		CAPIProviders:  []string{capiDockerProvider},
+	}
+}
+
 // NewBoostrapCluster Creates a new cluster manager for a local bootstrap cluster with the given
 // config, applying defaults where needed
 func NewBoostrapCluster(clusterConfig ClusterConfig) (ClusterLifeCycleManager, error) {
@@ -70,7 +77,7 @@ func NewBoostrapCluster(clusterConfig ClusterConfig) (ClusterLifeCycleManager, e
 }
 
 func setDefaults(c ClusterConfig) ClusterConfig {
-	defaultConfig := newDefaultConfig()
+	defaultConfig := NewClusterConfig()
 	actualConfig := ClusterConfig{
 		ClusterName:    c.ClusterName,
 		Type:           c.Type,
@@ -107,32 +114,4 @@ func validateConfig(config ClusterConfig) error {
 func unknownClusterTypeError(clusterType string) error {
 	return fmt.Errorf("unsupported cluster type %s - supported types are %v",
 		clusterType, publicSupportedClusterTypes)
-}
-
-func initCAPI(clcm ClusterLifeCycleManager) error {
-	config, err := createKubeConfigFile(clcm)
-	if err != nil {
-		return err
-	}
-	defer os2.RemoveTempFiles(zap.S(), config.Name())
-	capiclient, err := capiInitFunc("")
-	if err != nil {
-		return err
-	}
-	_, err = capiclient.Init(clusterapi.InitOptions{
-		Kubeconfig: clusterapi.Kubeconfig{
-			Path: config.Name(),
-		},
-		InfrastructureProviders: defaultCAPIProviders,
-	})
-	return err
-}
-
-func newDefaultConfig() ClusterConfig {
-	return ClusterConfig{
-		ClusterName:    bootstrapClusterName,
-		Type:           OCNEClusterType,
-		ContainerImage: getDefaultBoostrapImage(OCNEClusterType),
-		CAPIProviders:  []string{capiDockerProvider},
-	}
 }
